renamed round to roundMerkle and callbacks to round
