+++ proto/sign/round_merkle.go
@@ -10,6 +10,169 @@ import (
 "github.com/dedis/cothority/lib/coconet"
 )
 
+const FIRST_ROUND int = 1 // start counting rounds at 1
+
+type RoundMerkle struct {
+								   // Message created by root. It can be empty and it will make no difference. In
+								   // the case of a timestamp service however we need the timestamp generated by
+								   // the round for this round . It will be included in the challenge, and then
+								   // can be verified by the client
+	Msg            []byte
+	C              abstract.Secret // round lasting challenge
+	R              abstract.Secret // round lasting response
+
+	Log            SNLog           // round lasting log structure
+	HashedLog      []byte
+
+	R_hat          abstract.Secret // aggregate of responses
+
+	X_hat          abstract.Point  // aggregate of public keys
+
+	Commits        []*SigningMessage
+	Responses      []*SigningMessage
+
+								   // own big merkle subtree
+	MTRoot         hashid.HashId   // mt root for subtree, passed upwards
+	Leaves         []hashid.HashId // leaves used to build the merkle subtre
+	LeavesFrom     []string        // child names for leaves
+
+								   // mtRoot before adding HashedLog
+	LocalMTRoot    hashid.HashId
+
+								   // merkle tree roots of children in strict order
+	CMTRoots       []hashid.HashId
+	CMTRootNames   []string
+	Proofs         map[string]proof.Proof
+	Proof          []hashid.HashId
+	PubKey         abstract.Point
+	PrivKey        abstract.Secret
+	Name           string
+
+								   // round-lasting public keys of children servers that did not
+								   // respond to latest commit or respond phase, in subtree
+	ExceptionList  []abstract.Point
+								   // combined point commits of children servers in subtree
+	ChildV_hat     map[string]abstract.Point
+								   // combined public keys of children servers in subtree
+	ChildX_hat     map[string]abstract.Point
+								   // for internal verification purposes
+	ExceptionX_hat abstract.Point
+	ExceptionV_hat abstract.Point
+
+	BackLink       hashid.HashId
+	AccRound       []byte
+
+	Vote           *Vote
+	Suite          abstract.Suite
+
+	Children       map[string]coconet.Conn
+	Parent         string
+	View           int
+}
+
+type RoundType int
+
+const (
+	EmptyRT RoundType = iota
+	ViewChangeRT
+	AddRT
+	RemoveRT
+	ShutdownRT
+	NoOpRT
+	SigningRT
+)
+
+func NewRound(suite abstract.Suite) *RoundMerkle {
+	round := &RoundMerkle{}
+	round.Commits = make([]*SigningMessage, 0)
+	round.Responses = make([]*SigningMessage, 0)
+	round.ExceptionList = make([]abstract.Point, 0)
+	round.Suite = suite
+	round.Log.Suite = suite
+	return round
+}
+
+// Sets up a round according to the needs stated in the
+// Announcementmessage.
+func RoundSetup(sn *Node, view int, am *AnnouncementMessage) error {
+	sn.viewmu.Lock()
+	if sn.ChangingView && am.Vote != nil && am.Vote.Vcv == nil {
+		dbg.Lvl4(sn.Name(), "currently chaning view")
+		sn.viewmu.Unlock()
+		return ChangingViewError
+	}
+	sn.viewmu.Unlock()
+
+	sn.roundmu.Lock()
+	roundNbr := am.RoundNbr
+	if roundNbr <= sn.LastSeenRound {
+		sn.roundmu.Unlock()
+		return ErrPastRound
+	}
+
+	// make space for round type
+	if len(sn.RoundTypes) <= roundNbr {
+		sn.RoundTypes = append(sn.RoundTypes, make([]RoundType, max(len(sn.RoundTypes), roundNbr + 1))...)
+	}
+	if am.Vote == nil {
+		dbg.Lvl4(roundNbr, len(sn.RoundTypes))
+		sn.RoundTypes[roundNbr] = SigningRT
+	} else {
+		sn.RoundTypes[roundNbr] = RoundType(am.Vote.Type)
+	}
+	sn.roundmu.Unlock()
+
+	// set up commit and response channels for the new round
+	round := NewRound(sn.suite)
+	round.Vote = am.Vote
+	round.Children = sn.Children(view)
+	round.Parent = sn.Parent(view)
+	round.View = view
+	round.PubKey = sn.PubKey
+	round.PrivKey = sn.PrivKey
+	round.Name = sn.Name()
+	round.InitCommitCrypto()
+	sn.Rounds[roundNbr] = round
+
+	// update max seen round
+	sn.roundmu.Lock()
+	sn.LastSeenRound = max(sn.LastSeenRound, roundNbr)
+	sn.roundmu.Unlock()
+
+	// the root is the only node that keeps track of round # internally
+	if sn.IsRoot(view) {
+		sn.RoundsAsRoot += 1
+		// TODO: is sn.Round needed if we have LastSeenRound
+		sn.Round = roundNbr
+
+		// Create my back link to previous round
+		sn.SetBackLink(roundNbr)
+		// sn.SetAccountableRound(Round)
+	}
+	return nil
+}
+
+func (rt RoundType) String() string {
+	switch rt {
+	case EmptyRT:
+		return "empty"
+	case SigningRT:
+		return "signing"
+	case ViewChangeRT:
+		return "viewchange"
+	case AddRT:
+		return "add"
+	case RemoveRT:
+		return "remove"
+	case ShutdownRT:
+		return "shutdown"
+	case NoOpRT:
+		return "noop"
+	default:
+		return ""
+	}
+}
+
 /*
  * This is a module for the round-struct that does all the
  * calculation for a merkle-hash-tree.
@@ -17,7 +180,7 @@ import (
 
 // Create round lasting secret and commit point v and V
 // Initialize log structure for the round
-func (round *Round) InitCommitCrypto() {
+func (round *RoundMerkle) InitCommitCrypto() {
 	// generate secret and point commitment for this round
 	rand := round.Suite.Cipher([]byte(round.Name))
 	round.Log = SNLog{}
@@ -33,7 +196,7 @@ func (round *Round) InitCommitCrypto() {
 }
 
 // Adds a child-node to the Merkle-tree and updates the root-hashes
-func (round *Round) MerkleAddChildren() {
+func (round *RoundMerkle) MerkleAddChildren() {
 	// children commit roots
 	round.CMTRoots = make([]hashid.HashId, len(round.Leaves))
 	copy(round.CMTRoots, round.Leaves)
@@ -49,14 +212,14 @@ func (round *Round) MerkleAddChildren() {
 
 // Adds the local Merkle-tree root, usually from a stamper or
 // such
-func (round *Round) MerkleAddLocal(localMTroot hashid.HashId) {
+func (round *RoundMerkle) MerkleAddLocal(localMTroot hashid.HashId) {
 	// add own local mtroot to leaves
 	round.LocalMTRoot = localMTroot
 	round.Leaves = append(round.Leaves, round.LocalMTRoot)
 }
 
 // Hashes the log of the round-structure
-func (round *Round) MerkleHashLog() error {
+func (round *RoundMerkle) MerkleHashLog() error {
 	var err error
 
 	h := round.Suite.Hash()
@@ -70,7 +233,7 @@ func (round *Round) MerkleHashLog() error {
 }
 
 
-func (round *Round) ComputeCombinedMerkleRoot() {
+func (round *RoundMerkle) ComputeCombinedMerkleRoot() {
 	// add hash of whole log to leaves
 	round.Leaves = append(round.Leaves, round.HashedLog)
 
@@ -98,7 +261,7 @@ func (round *Round) ComputeCombinedMerkleRoot() {
 // Identify which proof corresponds to which leaf
 // Needed given that the leaves are sorted before passed to the function that create
 // the Merkle Tree and its Proofs
-func (round *Round) SeparateProofs(proofs []proof.Proof, leaves []hashid.HashId) {
+func (round *RoundMerkle) SeparateProofs(proofs []proof.Proof, leaves []hashid.HashId) {
 	// separate proofs for children servers mt roots
 	for i := 0; i < len(round.CMTRoots); i++ {
 		name := round.CMTRootNames[i]
@@ -119,7 +282,7 @@ func (round *Round) SeparateProofs(proofs []proof.Proof, leaves []hashid.HashId)
 	}
 }
 
-func (round *Round) InitResponseCrypto() {
+func (round *RoundMerkle) InitResponseCrypto() {
 	round.R = round.Suite.Secret()
 	round.R.Mul(round.PrivKey, round.C).Sub(round.Log.v, round.R)
 	// initialize sum of children's responses
@@ -129,7 +292,7 @@ func (round *Round) InitResponseCrypto() {
 // Create Merkle Proof for local client (timestamp server) and
 // store it in Node so that we can send it to the clients during
 // the SignatureBroadcast
-func (round *Round) StoreLocalMerkleProof(chm *ChallengeMessage) error {
+func (round *RoundMerkle) StoreLocalMerkleProof(chm *ChallengeMessage) error {
 	proofForClient := make(proof.Proof, len(chm.Proof))
 	copy(proofForClient, chm.Proof)
 
@@ -149,7 +312,7 @@ func (round *Round) StoreLocalMerkleProof(chm *ChallengeMessage) error {
 // Figure out which kids did not submit messages
 // Add default messages to messgs, one per missing child
 // as to make it easier to identify and add them to exception lists in one place
-func (round *Round) FillInWithDefaultMessages() []*SigningMessage {
+func (round *RoundMerkle) FillInWithDefaultMessages() []*SigningMessage {
 	children := round.Children
 
 	messgs := round.Responses
@@ -175,7 +338,7 @@ func (round *Round) FillInWithDefaultMessages() []*SigningMessage {
 }
 
 // Called by every node after receiving aggregate responses from descendants
-func (round *Round) VerifyResponses() error {
+func (round *RoundMerkle) VerifyResponses() error {
 
 	// Check that: base**r_hat * X_hat**c == V_hat
 	// Equivalent to base**(r+xc) == base**(v) == T in vanillaElGamal
@@ -211,7 +374,7 @@ func (round *Round) VerifyResponses() error {
 
 // Create Personalized Merkle Proofs for children servers
 // Send Personalized Merkle Proofs to children servers
-func (round *Round) SendChildrenChallengesProofs(chm *ChallengeMessage) error {
+func (round *RoundMerkle) SendChildrenChallengesProofs(chm *ChallengeMessage) error {
 	// proof from big root to our root will be sent to all children
 	baseProof := make(proof.Proof, len(chm.Proof))
 	copy(baseProof, chm.Proof)
@@ -236,7 +399,7 @@ func (round *Round) SendChildrenChallengesProofs(chm *ChallengeMessage) error {
 }
 
 // Send children challenges
-func (round *Round) SendChildrenChallenges(chm *ChallengeMessage) error {
+func (round *RoundMerkle) SendChildrenChallenges(chm *ChallengeMessage) error {
 	for _, child := range round.Children {
 		var messg coconet.BinaryMarshaler
 		messg = &SigningMessage{View: round.View, Type: Challenge, Chm: chm}
@@ -248,4 +411,33 @@ func (round *Round) SendChildrenChallenges(chm *ChallengeMessage) error {
 	}
 
 	return nil
+}
+
+
+// Adding-function for crypto-points that accepts nil
+func (r *RoundMerkle) Add(a abstract.Point, b abstract.Point) {
+	if a == nil {
+		a = r.Suite.Point().Null()
+	}
+	if b != nil {
+		a.Add(a, b)
+	}
+}
+
+// Substraction-function for crypto-points that accepts nil
+func (r *RoundMerkle) Sub(a abstract.Point, b abstract.Point) {
+	if a == nil {
+		a = r.Suite.Point().Null()
+	}
+	if b != nil {
+		a.Sub(a, b)
+	}
+}
+
+func (r *RoundMerkle) IsRoot() bool {
+	return r.Parent == ""
+}
+
+func (r *RoundMerkle) IsLeaf() bool {
+	return len(r.Children) == 0
 }
